Swap Combobox <-> ControlledCombobox and remove ComboboxRaw
